requirements.txt
gradio>=4.0.0
sentence-transformers>=2.2.0
faiss-cpu>=1.7.4
scikit-learn>=1.3.0
joblib>=1.3.0
pandas>=2.0.0
pyarrow>=14.0.0
numpy>=1.24.0
